requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Mail
Jinja2

python-dateutil
